backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
httpx==0.27.0
pyjwt[crypto]==2.8.0
supabase==2.6.0
pydantic>=2.7.0
pydantic-settings>=2.2.1
